requirements/base.txt
Django==5.0.6
asgiref==3.8.1
sqlparse==0.5.0
asgiref==3.8.1
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
sqlparse==0.5.0
python-decouple==3.8
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.3.1

